refactor game base files into /game
