+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:usesCleartextTraffic="true">
 
         <activity
-            android:name=".EndGameActivity"
+            android:name=".game.EndGameActivity"
             android:exported="false"
             android:parentActivityName=".MainActivity">
             <!-- The meta-data tag is required if you support API level 15 and lower -->
@@ -30,7 +30,7 @@
                 android:value=".MainActivity" />
         </activity>
         <activity
-            android:name=".GameActivity"
+            android:name=".game.GameActivity"
             android:launchMode= "singleInstance"
             android:exported="false" />
         <activity
