+++ app/src/main/java/com/github/fribourgsdp/radio/GameHandler.kt
@@ -1,5 +1,6 @@
 package com.github.fribourgsdp.radio
 
+import com.github.fribourgsdp.radio.game.GameView
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.EventListener
